[CI-1957]: add servicename support (#24960)
* [CI-1947]: fix functor

* [CI-1957]: add servicename support

+++ 930-delegate-tasks/src/main/java/io/harness/delegate/task/citasks/cik8handler/CIK8CleanupTaskHandler.java
@@ -53,12 +53,17 @@ public class CIK8CleanupTaskHandler implements CICleanupTaskHandler {
       try (
           KubernetesClient kubernetesClient = k8sConnectorHelper.createKubernetesClient(taskParams.getK8sConnector())) {
         boolean podsDeleted = deletePods(kubernetesClient, namespace, taskParams.getPodNameList());
-        boolean serviceDeleted = deleteServices(kubernetesClient, namespace, taskParams.getServiceNameList());
-        boolean secretsDeleted = deleteSecrets(
-            kubernetesClient, namespace, taskParams.getPodNameList(), taskParams.getCleanupContainerNames());
-        if (podsDeleted && serviceDeleted && secretsDeleted) {
-          return K8sTaskExecutionResponse.builder().commandExecutionStatus(CommandExecutionStatus.SUCCESS).build();
+        if (podsDeleted) {
+          boolean serviceDeleted = deleteServices(kubernetesClient, namespace, taskParams.getServiceNameList());
+          boolean secretsDeleted = deleteSecrets(
+              kubernetesClient, namespace, taskParams.getPodNameList(), taskParams.getCleanupContainerNames());
+          if (podsDeleted && serviceDeleted && secretsDeleted) {
+            return K8sTaskExecutionResponse.builder().commandExecutionStatus(CommandExecutionStatus.SUCCESS).build();
+          } else {
+            return K8sTaskExecutionResponse.builder().commandExecutionStatus(CommandExecutionStatus.FAILURE).build();
+          }
         } else {
+          log.error("Failed to delete pod {}", taskParams.getPodNameList());
           return K8sTaskExecutionResponse.builder().commandExecutionStatus(CommandExecutionStatus.FAILURE).build();
         }
       } catch (Exception ex) {
